requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
sentence-transformers==5.2.2
faiss-cpu==1.13.2
rank-bm25==0.2.2
transformers==5.1.0
torch==2.4.1
accelerate==1.12.0
openai==2.17.0
python-dotenv==1.0.0
